pyproject.toml
[project]
name = "llm-cli-tools-core"
version = "0.1.0"
description = "Core telemetry and utilities for LLM CLI tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "filelock>=3.12.0",
    "rich>=13.7.0",
]

[project.scripts]
llm-telemetry = "llm_cli_core.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.hatch.build.targets.wheel]
packages = ["src/llm_cli_core"]
